+++ CHANGELOG.rst
@@ -11,6 +11,8 @@ Change Log
 
 * The Zetterling MD pair potential: ``hoomd.md.pair.Zetterling``
   (`#2086 <https://github.com/glotzerlab/hoomd-blue/pull/2086>`__).
+* The YLZ anisotropic MD pair potential: ``hoomd.md.pair.aniso.YLZ``
+  (`#2110 <https://github.com/glotzerlab/hoomd-blue/pull/2110>`__).
 
 *Fixed*